+++ examples/plot.rs
@@ -6,7 +6,7 @@ use std::error::Error;
 extern crate robotics;
 use robotics::utils::state::GaussianStateStatic as GaussianState;
 
-pub fn ellipse_series(xy: Vector2<f32>, p_xy: Matrix2<f32>) -> Option<Vec<(f64, f64)>> {
+pub fn ellipse_series(xy: Vector2<f64>, p_xy: Matrix2<f64>) -> Option<Vec<(f64, f64)>> {
     let eigen = p_xy.symmetric_eigen();
     let eigenvectors = eigen.eigenvectors;
     let eigenvalues = eigen.eigenvalues;
@@ -15,13 +15,13 @@ pub fn ellipse_series(xy: Vector2<f32>, p_xy: Matrix2<f32>) -> Option<Vec<(f64, 
         (
             eigenvalues.x.sqrt(),
             eigenvalues.y.sqrt(),
-            f32::atan2(eigenvectors.m21, eigenvectors.m22),
+            f64::atan2(eigenvectors.m21, eigenvectors.m22),
         )
     } else {
         (
             eigenvalues.y.sqrt(),
             eigenvalues.x.sqrt(),
-            f32::atan2(eigenvectors.m22, eigenvectors.m21),
+            f64::atan2(eigenvectors.m22, eigenvectors.m21),
         )
     };
 
@@ -30,7 +30,7 @@ pub fn ellipse_series(xy: Vector2<f32>, p_xy: Matrix2<f32>) -> Option<Vec<(f64, 
         .clone_owned();
 
     let ellipse_points = (0..100)
-        .map(|x| x as f32 / 100.0 * std::f32::consts::TAU) // map [0..100] -> [0..2pi]
+        .map(|x| x as f64 / 100.0 * std::f64::consts::TAU) // map [0..100] -> [0..2pi]
         .map(|t| rot_mat * Vector2::new(a * t.cos(), b * t.sin()) + xy) //map [0..2pi] -> elipse points
         .map(|xy| (xy.x as f64, xy.y as f64))
         .collect();
@@ -43,8 +43,8 @@ pub struct History {
     pub x_true: Vec<(f64, f64)>,
     pub x_dr: Vec<(f64, f64)>,
     pub x_est: Vec<(f64, f64)>,
-    // pub gaussian_state: Vec<GaussianState<f32, Const<4>>>,
-    pub gaussian_state: Vec<GaussianState<f32, 4>>,
+    // pub gaussian_state: Vec<GaussianState<f64, Const<4>>>,
+    pub gaussian_state: Vec<GaussianState<f64, 4>>,
 }
 
 pub fn chart(
