+++ src/localization/extended_kalman_filter.rs
@@ -152,19 +152,19 @@ mod tests {
     #[test]
     fn ekf_runs() {
         // setup ukf
-        let q = Matrix4::<f32>::from_diagonal(&Vector4::new(0.1, 0.1, deg2rad(1.0), 1.0));
+        let q = Matrix4::<f64>::from_diagonal(&Vector4::new(0.1, 0.1, deg2rad(1.0), 1.0));
         let r = nalgebra::Matrix2::identity();
         let motion_model = Box::new(SimpleProblemMotionModel {});
         let measurement_model = Box::new(SimpleProblemMeasurementModel {});
-        let ekf = ExtendedKalmanFilter::<f32, 4, 2, 2>::new(q, r, measurement_model, motion_model);
+        let ekf = ExtendedKalmanFilter::<f64, 4, 2, 2>::new(q, r, measurement_model, motion_model);
 
         let dt = 0.1;
-        let u: Vector2<f32> = Default::default();
+        let u: Vector2<f64> = Default::default();
         let kalman_state = GaussianState {
-            x: Vector4::<f32>::new(0., 0., 0., 0.),
-            P: Matrix4::<f32>::identity(),
+            x: Vector4::<f64>::new(0., 0., 0., 0.),
+            P: Matrix4::<f64>::identity(),
         };
-        let z: Vector2<f32> = Default::default();
+        let z: Vector2<f64> = Default::default();
 
         ekf.estimate(&kalman_state, &u, &z, dt);
     }
